+++ spec/helper/expect_helpers.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+def expect_command(result, string)
+  expect(result.command.raw_command).to eq(
+    string
+  )
+end